% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{reverse_complement_dna}
\alias{reverse_complement_dna}
\title{Reverse-complement a DNA sequence}
\usage{
reverse_complement_dna(sequence)
}
\arguments{
\item{sequence}{Character scalar containing A, C, G, T, or IUPAC
ambiguous DNA bases.}
}
\value{
A character scalar with the reverse complement.
}
\description{
Reverse-complement a DNA sequence
}
\examples{
reverse_complement_dna("CCCTAACCCTAACCCTAACCCTAACCC")
}
